+++ src/utils/curriculum/slugs.test.ts
@@ -1,5 +1,6 @@
 import {
   createProgrammeSlug,
+  createTeacherProgrammeSlug,
   getKs4RedirectSlug,
   isValidSubjectPhaseSlug,
   KS4_EXAMBOARD_PREFERENCE,
@@ -120,7 +121,7 @@ describe("getKs4RedirectSlug", () => {
   });
 });
 
-describe("createProgrammeSlug", () => {
+describe("createProgrammeSlug/createTeacherProgrammeSlug", () => {
   test("unit data ks2 returns correct programme slug", () => {
     const unitData = {
       planned_number_of_lessons: 5,
@@ -156,6 +157,7 @@ describe("createProgrammeSlug", () => {
       state: "published",
     };
     expect(createProgrammeSlug(unitData)).toEqual("science-primary-ks2");
+    expect(createTeacherProgrammeSlug(unitData)).toEqual("science-primary-ks2");
   });
   test("unit data ks4 returns correct programme slug", () => {
     const unitData = {
@@ -192,6 +194,7 @@ describe("createProgrammeSlug", () => {
       state: "published",
     };
     expect(createProgrammeSlug(unitData)).toEqual("science-primary-ks4");
+    expect(createTeacherProgrammeSlug(unitData)).toEqual("science-primary-ks4");
   });
   test("unit data with exam board and tier returns the correct programme slug", () => {
     const unitData = {
@@ -230,6 +233,9 @@ describe("createProgrammeSlug", () => {
     expect(createProgrammeSlug(unitData, "aqa", "foundation")).toEqual(
       "combined-science-secondary-ks4-foundation-aqa",
     );
+    expect(createTeacherProgrammeSlug(unitData, "aqa", "foundation")).toEqual(
+      "combined-science-secondary-ks4-foundation-aqa",
+    );
   });
   test("unit data for ks3 returns the correct programme slug", () => {
     const unitData = {
@@ -266,6 +272,9 @@ describe("createProgrammeSlug", () => {
       state: "published",
     };
     expect(createProgrammeSlug(unitData, "aqa")).toEqual(
+      "combined-science-secondary-ks3",
+    );
+    expect(createTeacherProgrammeSlug(unitData, "aqa")).toEqual(
       "combined-science-secondary-ks3",
     );
   });
@@ -307,5 +316,50 @@ describe("createProgrammeSlug", () => {
     expect(createProgrammeSlug(unitData, "aqa")).toEqual(
       "combined-science-secondary-ks3",
     );
+    expect(createTeacherProgrammeSlug(unitData, "aqa")).toEqual(
+      "combined-science-secondary-ks3",
+    );
+  });
+
+  test("unit data with exam board and tier returns the correct programme slug", () => {
+    const unitData = {
+      planned_number_of_lessons: 5,
+      connection_future_unit_description: null,
+      connection_prior_unit_description: null,
+      connection_future_unit_title: null,
+      connection_prior_unit_title: null,
+      domain: null,
+      domain_id: null,
+      examboard: null,
+      examboard_slug: null,
+      keystage_slug: "ks4",
+      lessons: [],
+      order: 1,
+      phase: "Secondary",
+      phase_slug: "secondary",
+      slug: "cellular-respiration-and-atp",
+      subject: "Combined Science",
+      subject_parent: "Science",
+      subject_parent_slug: "science",
+      subject_slug: "combined-science",
+      tags: null,
+      subjectcategories: null,
+      threads: [],
+      tier: null,
+      tier_slug: null,
+      title: "Aerobic and anaerobic cellular respiration",
+      unit_options: [],
+      year: "11",
+      cycle: "1",
+      why_this_why_now: null,
+      description: null,
+      state: "published",
+    };
+    expect(createProgrammeSlug(unitData, "aqa", "foundation")).toEqual(
+      "combined-science-secondary-ks4-foundation-aqa",
+    );
+    expect(
+      createTeacherProgrammeSlug(unitData, "aqa", "foundation", "gcse"),
+    ).toEqual("combined-science-secondary-ks4-foundation-gcse-aqa");
   });
 });